Minor bug fix
